[Refactoring ⚙️] 댓글 버튼 컴포넌트 분리
