+++ src/pages/ProjectDetailPage/components/Comments/Comments.tsx
@@ -1,8 +1,8 @@
-import { Flex } from "@chakra-ui/react"
 import { Comment } from "api-models"
 
 import CommentsInput from "./CommentsInput"
 import CommentsList from "./CommentsList"
+import CommentsLayout from "./Layout/CommentsLayout"
 
 export interface CommentsProps {
   comments: Comment[]
@@ -10,18 +10,10 @@ export interface CommentsProps {
 
 const Comments = ({ comments }: CommentsProps) => {
   return (
-    <Flex
-      margin="0 auto"
-      w="50%"
-      maxW="128rem"
-      borderRadius="1rem"
-      border="1px"
-      borderColor="grey.400"
-      mt="10rem"
-      direction="column">
+    <CommentsLayout>
       <CommentsInput />
       <CommentsList comments={comments} />
-    </Flex>
+    </CommentsLayout>
   )
 }
