+++ src/pages/ProjectDetailPage/mocks/index.ts
@@ -1,4 +1,3 @@
-import { Comment } from "api-models"
 import { rest } from "msw"
 
 import { DUMMY_PROJECT_DETAIL } from "./mockData"
@@ -13,12 +12,10 @@ export const projectDetailHandlers = [
   //   return res(ctx.status(200))
   // }),
 
-  rest.delete("/api/v1/projects/:projectId/comments/:id", (req, res, ctx) => {
-    console.log(req.params.id)
-    const commentIdx = DUMMY_PROJECT_DETAIL.projects.comments.findIndex(
-      (comment: Comment) => Number(comment.id) === Number(req.params.id),
-    )
-    DUMMY_PROJECT_DETAIL.projects.comments.splice(commentIdx, 1)
-    return res(ctx.status(200))
-  }),
+  // rest.delete("/api/v1/projects/:projectId/comments/:id", (req, res, ctx) => {
+  //   const commentIdx = DUMMY_PROJECT_DETAIL.projects.comments.findIndex(
+  //     (comment: Comment) => Number(comment.id) === Number(req.params.id),
+  //   )
+  //   DUMMY_PROJECT_DETAIL.projects.comments.splice(commentIdx, 1)
+  //   return res(ctx.status(200))
 ]
